threat-analyzer-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
sentence-transformers
pandas
pypdf
python-dotenv
anthropic
requests
langchain
langchain-anthropic
sqlalchemy
pymysql
bcrypt
